rk is ScipyFit
