+++ src/probnum/diffeq/odefilter/init/__init__.py
@@ -2,5 +2,5 @@
 
 
 from ._autodiff import ForwardMode, ForwardModeJVP, ReverseMode, TaylorMode
-from ._runge_kutta import RungeKutta, RungeKuttaWithJacobian
+from ._scipy_fit import SciPyFit, SciPyFitWithJacobian
 from ._stack import Stack, StackWithJacobian
